+++ packages/frontend/src/api/useDialogByIdSubscription.ts
@@ -29,7 +29,6 @@ export const useDialogByIdSubscription = (dialogId: string | undefined, dialogTo
         if (updatedType && updatedType === DialogEventType.DialogDeleted) {
           // Redirect to inbox if the dialog was deleted
           navigate(PageRoutes.inbox);
-          void queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.DIALOGS] });
         } else if (updatedType && updatedType === DialogEventType.DialogUpdated) {
           void queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.DIALOG_BY_ID] });
           void queryClient.invalidateQueries({ queryKey: [QUERY_KEYS.DIALOGS] });