+++ packages/frontend/src/components/ProfileCheckbox/index.ts
@@ -1,1 +1,0 @@
-export { ProfileCheckbox } from './ProfileCheckbox.tsx';